+++ SwiftExample/ApiController.swift
@@ -13,13 +13,16 @@ protocol APIControllerProtocol {
 }
 
 class APIController: NSObject {
-    
+    let maximumResponseSize = 1024 * 1024
     var data: NSMutableData = NSMutableData()
-    
-    
+    var responseAccepted = false
+    var requestCompleted = false
     var delegate: APIControllerProtocol?
     
     func searchItunesFor(searchTerm: String) {
+        requestCompleted = false
+        responseAccepted = false
+        data = NSMutableData()
         let allowedCharacters = NSCharacterSet(charactersInString: "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789-._~")
 
         if let escapedSearchTerm = searchTerm.stringByAddingPercentEncodingWithAllowedCharacters(allowedCharacters) {
@@ -37,8 +40,35 @@ class APIController: NSObject {
     }
 
     func completeWithResults(results: NSDictionary) {
+        if requestCompleted {
+            return
+        }
+
+        requestCompleted = true
         delegate?.didRecieveAPIResults(results)
         self.data = NSMutableData()
+    }
+
+    func isAcceptableResponse(response: NSURLResponse) -> Bool {
+        guard let httpResponse = response as? NSHTTPURLResponse where
+            httpResponse.statusCode >= 200 && httpResponse.statusCode < 300 else {
+            return false
+        }
+
+        let contentLength = response.expectedContentLength
+        if contentLength > Int64(maximumResponseSize) {
+            return false
+        }
+
+        guard let mimeType = response.MIMEType?.lowercaseString else {
+            return false
+        }
+
+        return mimeType == "application/json" || mimeType == "text/javascript"
+    }
+
+    func canAppendResponseData(chunk: NSData) -> Bool {
+        return responseAccepted && chunk.length <= maximumResponseSize - data.length
     }
     
     
@@ -48,16 +78,30 @@ class APIController: NSObject {
     
     
     func connection(connection: NSURLConnection, didReceiveResponse response: NSURLResponse) {
-        // Recieved a new request, clear out the data object
         self.data = NSMutableData()
+        responseAccepted = isAcceptableResponse(response)
+        if !responseAccepted {
+            connection.cancel()
+            completeWithResults(NSDictionary())
+        }
     }
     
     func connection(connection: NSURLConnection, didReceiveData data: NSData) {
-        // Append the recieved chunk of data to our data object
+        if !canAppendResponseData(data) {
+            connection.cancel()
+            completeWithResults(NSDictionary())
+            return
+        }
+
         self.data.appendData(data)
     }
     
     func connectionDidFinishLoading(connection: NSURLConnection) {
+        if !responseAccepted {
+            completeWithResults(NSDictionary())
+            return
+        }
+
         // Request complete, self.data should now hold the resulting info
         // Convert the retrieved data in to an object through JSON deserialization
         do {